refactor: ipc connection
