+++ Cargo.toml
@@ -22,19 +22,11 @@ tokio = { version = "1", default-features = false, features = [
 tokio-tungstenite = "0.28"
 rand = "0.9"
 http = "1.3"
-base64 = "0.22"
 httparse = "1.10"
 log = "0.4"
 urlencoding = "2.1"
 ts-rs = "11.0"
 pin-project = "1.1"
-hyper = { version = "1.8" }
-bytes = "1.11"
-http-body-util = "0.1"
-crossbeam = "0.8"
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
 
 [target.'cfg(windows)'.dependencies]
 windows-sys = "0.61"
