+++ prek.toml
@@ -1,0 +1,39 @@
+fail_fast = true
+
+[[repos]]
+repo = "https://github.com/crate-ci/typos"
+rev = "v1.43.4"
+hooks = [
+  { id = "typos" }
+]
+
+[[repos]]
+repo = "local"
+hooks = [
+  {
+    id = "cargo-fmt",
+    name = "cargo fmt",
+    description = "Format files with rustfmt.",
+    entry = "cargo +nightly fmt  -- --check",
+    language = "system",
+    types = ["rust"]
+  },
+  {
+    id = "cargo-check",
+    name = "cargo check",
+    description = "Check the package for errors.",
+    entry = "cargo check --all",
+    language = "system",
+    types = ["rust"],
+    pass_filenames = false
+  },
+  {
+    id = "cargo-clippy",
+    name = "cargo clippy",
+    description = "Lint rust sources",
+    entry = "cargo clippy --all-targets --all-features --tests --benches -- -D warnings",
+    language = "system",
+    types = ["rust"],
+    pass_filenames = false
+  }
+]
