+++ src/commands.rs
@@ -234,7 +234,7 @@ pub(crate) async fn upgrade_geo(state: State<'_, RwLock<Mihomo>>) -> Result<()> 
 pub(crate) async fn ws_traffic(
     state: State<'_, RwLock<Mihomo>>,
     on_message: Channel<serde_json::Value>,
-) -> Result<ConnectionId> {
+) -> Result<WebSocketConnectionId> {
     state
         .read()
         .await
@@ -248,7 +248,7 @@ pub(crate) async fn ws_traffic(
 pub(crate) async fn ws_memory(
     state: State<'_, RwLock<Mihomo>>,
     on_message: Channel<serde_json::Value>,
-) -> Result<ConnectionId> {
+) -> Result<WebSocketConnectionId> {
     state
         .read()
         .await
@@ -262,7 +262,7 @@ pub(crate) async fn ws_memory(
 pub(crate) async fn ws_connections(
     state: State<'_, RwLock<Mihomo>>,
     on_message: Channel<serde_json::Value>,
-) -> Result<ConnectionId> {
+) -> Result<WebSocketConnectionId> {
     state
         .read()
         .await
@@ -277,7 +277,7 @@ pub(crate) async fn ws_logs(
     state: State<'_, RwLock<Mihomo>>,
     level: LogLevel,
     on_message: Channel<serde_json::Value>,
-) -> Result<ConnectionId> {
+) -> Result<WebSocketConnectionId> {
     state
         .read()
         .await
@@ -300,7 +300,7 @@ pub(crate) async fn ws_logs(
 #[command]
 pub(crate) async fn ws_disconnect(
     state: State<'_, RwLock<Mihomo>>,
-    id: ConnectionId,
+    id: WebSocketConnectionId,
     force_timeout: Option<u64>,
 ) -> Result<()> {
     state.read().await.disconnect(id, force_timeout).await
