+++ src/mihomo.rs
@@ -10,18 +10,19 @@ use reqwest::{Method, RequestBuilder};
 use serde_json::json;
 use tokio_tungstenite::{
     client_async, connect_async,
-    tungstenite::{Message, client::IntoClientRequest, protocol::CloseFrame as ProtocolCloseFrame},
+    tungstenite::{
+        Message, client::IntoClientRequest, handshake::client::generate_key, protocol::CloseFrame as ProtocolCloseFrame,
+    },
 };
 
 use crate::{
-    Error, IpcConnectionPool, Result,
-    ipc::LocalSocket,
+    Error, Result,
     models::{
-        BaseConfig, CloseFrame, ConnectionId, ConnectionManager, Connections, CoreUpdaterChannel, ErrorResponse,
-        Groups, LogLevel, MihomoVersion, Protocol, Proxies, Proxy, ProxyDelay, ProxyProvider, ProxyProviders,
-        RuleProviders, Rules, WebSocketMessage, WebSocketWriter,
+        BaseConfig, CloseFrame, ConnectionManager, Connections, CoreUpdaterChannel, ErrorResponse, Groups, LogLevel,
+        MihomoVersion, Protocol, Proxies, Proxy, ProxyDelay, ProxyProvider, ProxyProviders, RuleProviders, Rules,
+        WebSocketConnectionId, WebSocketMessage, WebSocketWriter,
     },
-    ret_failed_resp, utils,
+    ret_failed_resp,
 };
 
 const DEFAULT_REQUEST_TIMEOUT: Duration = Duration::from_secs(5);
@@ -52,14 +53,6 @@ impl Mihomo {
     #[inline]
     pub fn update_secret(&mut self, secret: Option<String>) {
         self.secret = secret;
-    }
-
-    #[inline]
-    pub fn update_socket_path<S: Into<String>>(&mut self, socket_path: S) -> Result<()> {
-        self.socket_path = Some(socket_path.into());
-        let pool = IpcConnectionPool::global()?;
-        pool.clear_pool();
-        Ok(())
     }
 
     #[inline]
@@ -100,8 +93,34 @@ impl Mihomo {
     fn build_request(&self, method: Method, suffix_url: &str) -> Result<RequestBuilder> {
         let url = self.get_req_url(suffix_url)?;
         let headers = self.get_req_headers()?;
-        let client = reqwest::ClientBuilder::new().build()?;
-        let req = match method {
+        let client = match self.protocol {
+            Protocol::Http => reqwest::ClientBuilder::new().build()?,
+            Protocol::LocalSocket => {
+                let Some(socket_path) = &self.socket_path else {
+                    log::error!("missing socket path parameter");
+                    return Err(Error::Io(std::io::Error::new(
+                        std::io::ErrorKind::InvalidInput,
+                        "missing socket path".to_string(),
+                    )));
+                };
+                #[cfg(windows)]
+                {
+                    reqwest::ClientBuilder::new()
+                        .windows_named_pipe(socket_path.clone())
+                        .timeout(DEFAULT_REQUEST_TIMEOUT)
+                        .build()?
+                }
+                #[cfg(unix)]
+                {
+                    reqwest::ClientBuilder::new()
+                        .unix_socket(socket_path.clone())
+                        .timeout(DEFAULT_REQUEST_TIMEOUT)
+                        .build()?
+                }
+            }
+        };
+
+        match method {
             Method::POST => Ok(client.post(url).headers(headers)),
             Method::GET => Ok(client.get(url).headers(headers)),
             Method::PUT => Ok(client.put(url).headers(headers)),
@@ -111,26 +130,6 @@ impl Mihomo {
                 let method_str = method.as_str().to_string();
                 log::error!("method not supported: {method_str}");
                 Err(Error::MethodNotSupported(method_str))
-            }
-        };
-        // 在此设置 timeout，以供构建 local socket 连接时，获取到 timeout 属性
-        Ok(req?.timeout(DEFAULT_REQUEST_TIMEOUT))
-    }
-
-    async fn send_by_protocol(&self, client: RequestBuilder) -> Result<reqwest::Response> {
-        match self.protocol {
-            Protocol::Http => client.send().await.map_err(Error::Reqwest),
-            Protocol::LocalSocket => {
-                if let Some(socket_path) = self.socket_path.as_ref() {
-                    log::debug!("send to local socket: {socket_path}");
-                    client.send_by_local_socket(socket_path).await
-                } else {
-                    log::error!("missing socket path parameter");
-                    Err(Error::Io(std::io::Error::new(
-                        std::io::ErrorKind::InvalidInput,
-                        "missing socket path".to_string(),
-                    )))
-                }
             }
         }
     }
@@ -157,7 +156,7 @@ impl Mihomo {
     }
 
     /// 连接 WebSocket
-    async fn connect<F>(&self, url: String, on_message: F) -> Result<ConnectionId>
+    async fn connect<F>(&self, url: String, on_message: F) -> Result<WebSocketConnectionId>
     where
         F: Fn(serde_json::Value) + Send + 'static,
     {
@@ -227,12 +226,12 @@ impl Mihomo {
             Protocol::LocalSocket => {
                 if let Some(socket_path) = self.socket_path.as_ref() {
                     log::debug!("starting connect to websocket by using local socket: {socket_path}");
-                    let stream = crate::ipc::connect_to_socket(socket_path).await?;
+                    let stream = crate::wrap_stream::connect_to_socket(socket_path).await?;
 
                     let request = Request::builder()
                         .uri(url)
                         .header(HOST, "clash-verge")
-                        .header(SEC_WEBSOCKET_KEY, utils::generate_websocket_key())
+                        .header(SEC_WEBSOCKET_KEY, generate_key())
                         .header(CONNECTION, "Upgrade")
                         .header(UPGRADE, "websocket")
                         .header(SEC_WEBSOCKET_VERSION, "13")
@@ -278,7 +277,7 @@ impl Mihomo {
     }
 
     /// 向指定 WebSocket 连接发送消息 (暂无使用该方法的地方)
-    async fn send(&self, id: ConnectionId, message: WebSocketMessage) -> Result<()> {
+    async fn send(&self, id: WebSocketConnectionId, message: WebSocketMessage) -> Result<()> {
         let manager = Arc::clone(&self.connection_manager);
         let mut manager = manager.0.write().await;
         if let Some(writer) = manager.get_mut(&id) {
@@ -296,12 +295,12 @@ impl Mihomo {
             Ok(())
         } else {
             log::error!("connection not found: {id}");
-            Err(Error::ConnectionNotFound(id))
+            Err(Error::WebSocketConnectionNotFound(id))
         }
     }
 
     /// 取消 WebSocket 连接
-    pub async fn disconnect(&self, id: ConnectionId, force_timeout: Option<u64>) -> Result<()> {
+    pub async fn disconnect(&self, id: WebSocketConnectionId, force_timeout: Option<u64>) -> Result<()> {
         log::debug!("disconnecting connection: {id}");
         let mut manager = self.connection_manager.0.write().await;
         if let Some(writer) = manager.get_mut(&id) {
@@ -322,7 +321,7 @@ impl Mihomo {
             Ok(())
         } else {
             log::error!("connection not found: {id}");
-            Err(Error::ConnectionNotFound(id))
+            Err(Error::WebSocketConnectionNotFound(id))
         }
     }
 
@@ -340,7 +339,7 @@ impl Mihomo {
     // |                     Mihomo API                     |
     // ------------------------------------------------------
     /// WebSocket: Mihomo 流量数据
-    pub async fn ws_traffic<F>(&self, on_message: F) -> Result<ConnectionId>
+    pub async fn ws_traffic<F>(&self, on_message: F) -> Result<WebSocketConnectionId>
     where
         F: Fn(serde_json::Value) + Send + 'static,
     {
@@ -350,7 +349,7 @@ impl Mihomo {
     }
 
     /// WebSocket: Mihomo 内存使用数据
-    pub async fn ws_memory<F>(&self, on_message: F) -> Result<ConnectionId>
+    pub async fn ws_memory<F>(&self, on_message: F) -> Result<WebSocketConnectionId>
     where
         F: Fn(serde_json::Value) + Send + 'static,
     {
@@ -360,7 +359,7 @@ impl Mihomo {
     }
 
     /// WebSocket: Mihomo 连接信息数据
-    pub async fn ws_connections<F>(&self, on_message: F) -> Result<ConnectionId>
+    pub async fn ws_connections<F>(&self, on_message: F) -> Result<WebSocketConnectionId>
     where
         F: Fn(serde_json::Value) + Send + 'static,
     {
@@ -370,7 +369,7 @@ impl Mihomo {
     }
 
     /// WebSocket: Mihomo 日志数据
-    pub async fn ws_logs<F>(&self, level: LogLevel, on_message: F) -> Result<ConnectionId>
+    pub async fn ws_logs<F>(&self, level: LogLevel, on_message: F) -> Result<WebSocketConnectionId>
     where
         F: Fn(serde_json::Value) + Send + 'static,
     {
@@ -388,8 +387,7 @@ impl Mihomo {
     // clash api
     /// 获取 Mihomo 版本信息
     pub async fn get_version(&self) -> Result<MihomoVersion> {
-        let client = self.build_request(Method::GET, "/version")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/version")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get mihomo version failed, {}", e),
@@ -402,8 +400,7 @@ impl Mihomo {
 
     /// 清理 FakeIP 缓存
     pub async fn flush_fakeip(&self) -> Result<()> {
-        let client = self.build_request(Method::POST, "/cache/fakeip/flush")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::POST, "/cache/fakeip/flush")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("flush fakeip cache failed, {}", e),
@@ -416,8 +413,7 @@ impl Mihomo {
 
     /// 清理 DNS 缓存
     pub async fn flush_dns(&self) -> Result<()> {
-        let client = self.build_request(Method::POST, "/cache/dns/flush")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::POST, "/cache/dns/flush")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -430,8 +426,7 @@ impl Mihomo {
 
     /// 获取全部连接信息
     pub async fn get_connections(&self) -> Result<Connections> {
-        let client = self.build_request(Method::GET, "/connections")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/connections")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get all connections failed, {}", e),
@@ -444,8 +439,7 @@ impl Mihomo {
 
     /// 关闭全部连接
     pub async fn close_all_connections(&self) -> Result<()> {
-        let client = self.build_request(Method::DELETE, "/connections")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::DELETE, "/connections")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("close all connections failed, {}", e),
@@ -458,8 +452,10 @@ impl Mihomo {
 
     /// 关闭指定 ID 的连接
     pub async fn close_connection(&self, connection_id: &str) -> Result<()> {
-        let client = self.build_request(Method::DELETE, &format!("/connections/{connection_id}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::DELETE, &format!("/connections/{connection_id}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -472,8 +468,7 @@ impl Mihomo {
 
     /// 获取所有的代理组
     pub async fn get_groups(&self) -> Result<Groups> {
-        let client = self.build_request(Method::GET, "/group")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/group")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -487,8 +482,10 @@ impl Mihomo {
     /// 获取指定名称的代理组
     pub async fn get_group_by_name(&self, group_name: &str) -> Result<Proxy> {
         let group_name_encode = urlencoding::encode(group_name);
-        let client = self.build_request(Method::GET, &format!("/group/{group_name_encode}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::GET, &format!("/group/{group_name_encode}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get group[{}] failed, {}", group_name, e),
@@ -505,8 +502,11 @@ impl Mihomo {
         let test_url = urlencoding::encode(test_url);
         let suffix_url = format!("/group/{group_name_encode}/delay?url={test_url}&timeout={timeout}");
         let req_timeout = Duration::from_millis(timeout as u64) + DEFAULT_REQUEST_TIMEOUT;
-        let client = self.build_request(Method::GET, &suffix_url)?.timeout(req_timeout);
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::GET, &suffix_url)?
+            .timeout(req_timeout)
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("delay group[{}] failed, {}", group_name, e),
@@ -519,8 +519,7 @@ impl Mihomo {
 
     /// 获取代理提供者信息
     pub async fn get_proxy_providers(&self) -> Result<ProxyProviders> {
-        let client = self.build_request(Method::GET, "/providers/proxies")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/providers/proxies")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get all proxy providers failed, {}", e),
@@ -534,8 +533,10 @@ impl Mihomo {
     /// 获取指定代理提供者信息
     pub async fn get_proxy_provider_by_name(&self, provider_name: &str) -> Result<ProxyProvider> {
         let provider_name_encode = urlencoding::encode(provider_name);
-        let client = self.build_request(Method::GET, &format!("/providers/proxies/{provider_name_encode}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::GET, &format!("/providers/proxies/{provider_name_encode}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get proxy provider[{}] failed, {}", provider_name, e),
@@ -549,8 +550,10 @@ impl Mihomo {
     /// 更新指定代理提供者信息
     pub async fn update_proxy_provider(&self, provider_name: &str) -> Result<()> {
         let provider_name_encode = urlencoding::encode(provider_name);
-        let client = self.build_request(Method::PUT, &format!("/providers/proxies/{provider_name_encode}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::PUT, &format!("/providers/proxies/{provider_name_encode}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("update proxy provider[{}] failed, {}", provider_name, e),
@@ -564,9 +567,13 @@ impl Mihomo {
     /// 对指定代理提供者进行健康检查
     pub async fn healthcheck_proxy_provider(&self, provider_name: &str) -> Result<()> {
         let provider_name_encode = urlencoding::encode(provider_name);
-        let suffix_url = format!("/providers/proxies/{provider_name_encode}/healthcheck");
-        let client = self.build_request(Method::GET, &suffix_url)?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(
+                Method::GET,
+                &format!("/providers/proxies/{provider_name_encode}/healthcheck"),
+            )?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("healthcheck proxy provider[{}] failed, {}", provider_name, e),
@@ -587,13 +594,16 @@ impl Mihomo {
     ) -> Result<ProxyDelay> {
         let provider_name_encode = urlencoding::encode(provider_name);
         let proxy_name_encode = urlencoding::encode(proxy_name);
-        let suffix_url = format!("/providers/proxies/{provider_name_encode}/{proxy_name_encode}/healthcheck");
         let req_timeout = Duration::from_millis(timeout as u64) + DEFAULT_REQUEST_TIMEOUT;
-        let client = self
-            .build_request(Method::GET, &suffix_url)?
+        let response = self
+            .build_request(
+                Method::GET,
+                &format!("/providers/proxies/{provider_name_encode}/{proxy_name_encode}/healthcheck"),
+            )?
             .query(&[("url", test_url), ("timeout", &timeout.to_string())])
-            .timeout(req_timeout);
-        let response = self.send_by_protocol(client).await?;
+            .timeout(req_timeout)
+            .send()
+            .await?;
         if !response.status().is_success() {
             // maybe proxy delay is timeout response, try parse it.
             match response.json::<ErrorResponse>().await {
@@ -611,8 +621,7 @@ impl Mihomo {
 
     /// 获取所有代理信息
     pub async fn get_proxies(&self) -> Result<Proxies> {
-        let client = self.build_request(Method::GET, "/proxies")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/proxies")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -626,8 +635,10 @@ impl Mihomo {
     /// 获取指定代理信息
     pub async fn get_proxy_by_name(&self, proxy_name: &str) -> Result<Proxy> {
         let proxy_name_encode = urlencoding::encode(proxy_name);
-        let client = self.build_request(Method::GET, &format!("/proxies/{proxy_name_encode}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::GET, &format!("/proxies/{proxy_name_encode}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get proxy[{}] failed, {}", proxy_name, e),
@@ -643,11 +654,11 @@ impl Mihomo {
     /// 一般为指定代理组下使用指定的代理节点 【代理组/节点】
     pub async fn select_node_for_group(&self, group_name: &str, node: &str) -> Result<()> {
         let group_name_encode = urlencoding::encode(group_name);
-        let body = json!({ "name": node });
-        let client = self
+        let response = self
             .build_request(Method::PUT, &format!("/proxies/{group_name_encode}"))?
-            .json(&body);
-        let response = self.send_by_protocol(client).await?;
+            .json(&json!({ "name": node }))
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("select node[{}] for group[{}] failed, {}", node, group_name, e),
@@ -663,8 +674,10 @@ impl Mihomo {
     /// 一般用于自动选择的代理组（例如：URLTest 类型的代理组）下的节点
     pub async fn unfixed_proxy(&self, group_name: &str) -> Result<()> {
         let group_name_encode = urlencoding::encode(group_name);
-        let client = self.build_request(Method::DELETE, &format!("/proxies/{group_name_encode}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::DELETE, &format!("/proxies/{group_name_encode}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("unfixed group[{}] failed, {}", group_name, e),
@@ -680,13 +693,13 @@ impl Mihomo {
     /// 一般用于代理节点的延迟测试，也可传代理组名称（只会测试代理组下选中的代理节点）
     pub async fn delay_proxy_by_name(&self, proxy_name: &str, test_url: &str, timeout: u32) -> Result<ProxyDelay> {
         let proxy_name_encode = urlencoding::encode(proxy_name);
-        let suffix_url = format!("/proxies/{proxy_name_encode}/delay");
         let req_timeout = Duration::from_millis(timeout as u64) + DEFAULT_REQUEST_TIMEOUT;
-        let client = self
-            .build_request(Method::GET, &suffix_url)?
+        let response = self
+            .build_request(Method::GET, &format!("/proxies/{proxy_name_encode}/delay"))?
             .query(&[("timeout", &timeout.to_string()), ("url", &test_url.to_string())])
-            .timeout(req_timeout);
-        let response = self.send_by_protocol(client).await?;
+            .timeout(req_timeout)
+            .send()
+            .await?;
         if !response.status().is_success() {
             match response.json::<ErrorResponse>().await {
                 Ok(err_res) => {
@@ -707,8 +720,7 @@ impl Mihomo {
 
     /// 获取所有规则信息
     pub async fn get_rules(&self) -> Result<Rules> {
-        let client = self.build_request(Method::GET, "/rules")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/rules")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -721,8 +733,7 @@ impl Mihomo {
 
     /// 获取所有规则提供者信息
     pub async fn get_rule_providers(&self) -> Result<RuleProviders> {
-        let client = self.build_request(Method::GET, "/providers/rules")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/providers/rules")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("get all rule providers failed, {}", e),
@@ -736,8 +747,10 @@ impl Mihomo {
     /// 更新规则提供者信息
     pub async fn update_rule_provider(&self, provider_name: &str) -> Result<()> {
         let provider_name_encode = urlencoding::encode(provider_name);
-        let client = self.build_request(Method::PUT, &format!("/providers/rules/{provider_name_encode}"))?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self
+            .build_request(Method::PUT, &format!("/providers/rules/{provider_name_encode}"))?
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("update rule provider[{}] failed, {}", provider_name, e),
@@ -750,8 +763,7 @@ impl Mihomo {
 
     /// 获取基础配置
     pub async fn get_base_config(&self) -> Result<BaseConfig> {
-        let client = self.build_request(Method::GET, "/configs")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::GET, "/configs")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -762,21 +774,17 @@ impl Mihomo {
         Ok(response.json::<BaseConfig>().await?)
     }
 
-    /// 重新加载配置
+    /// 重新加载配置 (默认 60 秒超时)
+    ///
+    /// 如果配置文件中包含了很多 provider，则需等待 provider 下载完成 (如果网络不好则导致此方法耗时)
     pub async fn reload_config(&self, force: bool, config_path: &str) -> Result<()> {
-        let body = json!({ "path": config_path });
-        let client = self
+        let response = self
             .build_request(Method::PUT, "/configs")?
             .timeout(Duration::from_secs(60))
             .query(&[("force", force)])
-            .json(&body);
-        let response_result = self.send_by_protocol(client).await;
-        if matches!(self.protocol, Protocol::LocalSocket)
-            && let Ok(pool) = IpcConnectionPool::global()
-        {
-            pool.clear_pool();
-        }
-        let response = response_result?;
+            .json(&json!({ "path": config_path }))
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("reload base config failed, {}", e),
@@ -789,8 +797,7 @@ impl Mihomo {
 
     /// 更新基础配置
     pub async fn patch_base_config<D: serde::Serialize + Clone + Sync>(&self, data: &D) -> Result<()> {
-        let client = { self.build_request(Method::PATCH, "/configs")?.json(&data) };
-        let response = { self.send_by_protocol(client).await? };
+        let response = self.build_request(Method::PATCH, "/configs")?.json(data).send().await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("patch base config failed, {}", e),
@@ -803,10 +810,11 @@ impl Mihomo {
 
     /// 更新 Geo, 同 [`upgrade_geo`](crate::mihomo::Mihomo::upgrade_geo)
     pub async fn update_geo(&self) -> Result<()> {
-        let client = self
+        let response = self
             .build_request(Method::POST, "/configs/geo")?
-            .timeout(Duration::from_secs(60));
-        let response = self.send_by_protocol(client).await?;
+            .timeout(Duration::from_secs(60))
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("update geo database failed, {}", e),
@@ -819,8 +827,7 @@ impl Mihomo {
 
     /// 重启核心
     pub async fn restart(&self) -> Result<()> {
-        let client = self.build_request(Method::POST, "/restart")?;
-        let response = self.send_by_protocol(client).await?;
+        let response = self.build_request(Method::POST, "/restart")?.send().await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -833,11 +840,12 @@ impl Mihomo {
 
     /// 升级核心
     pub async fn upgrade_core(&self, channel: CoreUpdaterChannel, force: bool) -> Result<()> {
-        let client = self
+        let response = self
             .build_request(Method::POST, "/upgrade")?
             .timeout(Duration::from_secs(60))
-            .query(&[("channel", &channel.to_string()), ("force", &force.to_string())]);
-        let response = self.send_by_protocol(client).await?;
+            .query(&[("channel", &channel.to_string()), ("force", &force.to_string())])
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("upgrade core failed, {}", e),
@@ -857,10 +865,11 @@ impl Mihomo {
 
     /// 更新 UI
     pub async fn upgrade_ui(&self) -> Result<()> {
-        let client = self
+        let response = self
             .build_request(Method::POST, "/upgrade/ui")?
-            .timeout(Duration::from_secs(60));
-        let response = self.send_by_protocol(client).await?;
+            .timeout(Duration::from_secs(60))
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response
                 .json::<ErrorResponse>()
@@ -873,10 +882,11 @@ impl Mihomo {
 
     /// 更新 Geo
     pub async fn upgrade_geo(&self) -> Result<()> {
-        let client = self
+        let response = self
             .build_request(Method::POST, "/upgrade/geo")?
-            .timeout(Duration::from_secs(60));
-        let response = self.send_by_protocol(client).await?;
+            .timeout(Duration::from_secs(60))
+            .send()
+            .await?;
         if !response.status().is_success() {
             let err_msg = response.json::<ErrorResponse>().await.map_or_else(
                 |e| format!("upgrade geo database failed, {}", e),
