+++ src/models.rs
@@ -6,7 +6,7 @@ use tokio::{net::TcpStream, sync::RwLock};
 use tokio_tungstenite::{MaybeTlsStream, WebSocketStream, tungstenite::Message};
 use ts_rs::TS;
 
-use crate::ipc::WrapStream;
+use crate::wrap_stream::WrapStream;
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
@@ -892,7 +892,7 @@ pub enum WebSocketMessage {
     Close(Option<CloseFrame>),
 }
 
-pub type ConnectionId = u32;
+pub type WebSocketConnectionId = u32;
 pub enum WebSocketWriter {
     TcpStreamWriter(SplitSink<WebSocketStream<MaybeTlsStream<TcpStream>>, Message>),
     SocketStreamWriter(SplitSink<WebSocketStream<WrapStream>, Message>),
@@ -913,4 +913,4 @@ impl WebSocketWriter {
 }
 
 #[derive(Default)]
-pub struct ConnectionManager(pub RwLock<HashMap<ConnectionId, WebSocketWriter>>);
+pub struct ConnectionManager(pub RwLock<HashMap<WebSocketConnectionId, WebSocketWriter>>);
